Separate habit storage tests with notifications
